pyproject.toml
[project]
name = "multi-md-to-pdf"
version = "0.1.0"
description = "Recursively stitch Markdown files into a single PDF"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Carlos" }]
dependencies = [
  "markdown-it-py>=3.0.0",
  "weasyprint>=62.0",
  "typer>=0.12.0",
  "rich>=13.7.0"
]

[project.scripts]
multi-md-to-pdf = "multi_md_to_pdf.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

